Add new InputElement component in the UI package
A new component called InputElement has been implemented in the UI package. The InputElement component is a versatile input field that supports different types of inputs as specified by the 'type' prop. The component is based on the TextField component from Material UI and includes relevant prop types and onchange function.