Downgrade deploy-pages action version in CI workflow
The version of 'deploy-pages' action being used in the CI workflow has been downgraded from v4 to v3. This was necessary due to dependencies or features that may not have been supported or functional in the recently released version. The change was executed in the CI workflow file located in .github/workflows/.