Update project license from CC-BY-NC-4.0 to AGPL-3.0
The project's license has been changed from Creative Commons Attribution-NonCommercial 4.0 International (CC-BY-NC-4.0) to GNU Affero General Public License v3.0 (AGPL-3.0). This impacts the rights for sharing, adapting, and using the project, as defined in the README and reflected in the package.json file. For full license details, refer to the AGPL-3.0 license page linked in the README.