Add background customization to application
The update introduces a new feature that allows users to customize the background of the application. This includes setting the background color, size, and position. The changes also involved exporting the new PositionSelector component for use in the application. Additionally, the code exports newly added properties related to the background in the package.json file of the UI package.