Feature/metadata (#26)
* Add Google site verification and OpenGraph meta tag

The changes introduced add Google site verification and OpenGraph image meta tag to the head section of the root layout. This is essential for SEO and social media sharing as it helps Google verify the site, and provides an image for OpenGraph to use when the site's URL is shared.

* Add metadata and remove redundant tags in layout component

This commit extends the 'metadata' object, and subsequently removes duplicate meta information from the head tag in the layout.tsx component file. The changes include the addition of a base URL, openGraph, keywords, and verification properties to the metadata.